+++ components/ChatList/CreateChatButton/CreateChatButton.module.scss
@@ -1,5 +1,5 @@
 @import '@/styles/index.scss';
-@import '../../styles/_variables';
+@import '../../../styles/_variables';
 
 .chatPlusBtn {
   position: fixed;
